refactor: rename routes and constants for index results
- Renamed constants for index results routes in `constants.ts`.
- Updated test files to reflect the new route names.
- Renamed route handler files to match the new route names.
- Adjusted import statements and references to the renamed routes and constants.
