+++ x-pack/plugins/ecs_data_quality_dashboard/common/constants.ts
@@ -13,7 +13,7 @@ export const GET_INDEX_STATS = `${BASE_PATH}/stats/{pattern}`;
 export const GET_INDEX_MAPPINGS = `${BASE_PATH}/mappings/{pattern}`;
 export const GET_UNALLOWED_FIELD_VALUES = `${BASE_PATH}/unallowed_field_values`;
 export const GET_ILM_EXPLAIN = `${BASE_PATH}/ilm_explain/{pattern}`;
-export const RESULTS_ROUTE_PATH = `${BASE_PATH}/results`;
-export const RESULTS_INDICES_LATEST_ROUTE_PATH = `${BASE_PATH}/results/indices_latest/{pattern}`;
-export const RESULTS_INDICES_PATTERN_ROUTE_PATH = `${BASE_PATH}/results/indices/{pattern}`;
+export const POST_INDEX_RESULTS = `${BASE_PATH}/results`;
+export const GET_INDEX_RESULTS = `${BASE_PATH}/results/{pattern}`;
+export const GET_INDEX_RESULTS_LATEST = `${BASE_PATH}/results_latest/{pattern}`;
 export const INTERNAL_API_VERSION = '1';
