+++ x-pack/plugins/ecs_data_quality_dashboard/server/routes/results/post_index_results.ts
@@ -7,7 +7,7 @@
 
 import type { IRouter, Logger } from '@kbn/core/server';
 
-import { RESULTS_ROUTE_PATH, INTERNAL_API_VERSION } from '../../../common/constants';
+import { POST_INDEX_RESULTS, INTERNAL_API_VERSION } from '../../../common/constants';
 import { buildResponse } from '../../lib/build_response';
 import { buildRouteValidation } from '../../schemas/common';
 import { PostResultBody } from '../../schemas/result';
@@ -16,13 +16,13 @@ import type { DataQualityDashboardRequestHandlerContext } from '../../types';
 import { checkIndicesPrivileges } from './privileges';
 import { API_RESULTS_INDEX_NOT_AVAILABLE } from './translations';
 
-export const postResultsRoute = (
+export const postIndexResultsRoute = (
   router: IRouter<DataQualityDashboardRequestHandlerContext>,
   logger: Logger
 ) => {
   router.versioned
     .post({
-      path: RESULTS_ROUTE_PATH,
+      path: POST_INDEX_RESULTS,
       access: 'internal',
       options: { tags: ['access:securitySolution'] },
     })
